+++ src/main/java/com/roadtax/repositories/OwnerRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.rest.core.annotation.RepositoryRestResource;
 
 @RepositoryRestResource(collectionResourceRel = "owner", path="owner")
-public interface OwnerRepository extends JpaRepository<Owner, Long>
+public interface OwnerRepository extends JpaRepository<Owner, Integer>
 {
 }
